pyproject.toml
[tool.poetry]
name = "pilotlog"
version = "0.1.0"
description = "Import/Export Django App"
authors = ["Bahadır Özkan"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "4.1.0"
djangorestframework = "^3.14.0"
psycopg2 = "2.9.9"
django-import-export = "3.3.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-django",
    "ruff==0.1.13",
]

[tool.ruff]
line-length = 79
fix = true

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    'static',
    'migrations',
    '__pycache__',
    '.pytest_cache',
    '__init__.py',
]